+++ handpose/src/handpose_test.ts
@@ -20,11 +20,10 @@ import * as tf from '@tensorflow/tfjs-core';
 import {describeWithFlags, NODE_ENVS} from '@tensorflow/tfjs-core/dist/jasmine_util';
 
 import * as handpose from './index';
-import {HandPose} from './pipeline';
 import {stubbedImageVals} from './test_util';
 
 describeWithFlags('Handpose', NODE_ENVS, () => {
-  let model: HandPose;
+  let model: handpose.HandPose;
   beforeAll(async () => {
     model = await handpose.load({detectionConfidence: 0});
   });
